use of the longListSeparator
This is an example of how you might use the long list separator.  You could really use it anywhere.  As an example, I used it in a table where I had a long list of three-leter codes that looked like vomit when printed out.  This organizes them nicely into a semi-regular list with little bullets to break up the monotony.